catSite/requirements.txt
asgiref==3.8.1
dj-config-url==0.1.1
Django==5.1.6
django-appconf==1.1.0
django-image-cropping==1.7
easy-thumbnails==2.10
environ==1.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
whitenoise==6.9.0
